move translation back to root
