+++ translation/feature_extraction_module.py
@@ -11,6 +11,7 @@ from pathlib import Path
 from einops import rearrange
 from typing import Any, Generator, List, Union, Tuple
 
+from omegaconf import MISSING
 import pandas as pd
 import torch
 from torch import nn
@@ -26,27 +27,63 @@ from stopes.core import Requirements, StopesModule
 @dataclass
 class LauncherConfig:
     cluster: str = "slurm"
+    """Cluster to use by stopes launcher, e.g. 'slurm' or 'local'"""
+
     partition: str = "gpu"
+    """Partition to use in slurm environment"""
+
     max_jobarray_jobs: int = 128
+    """Maximum number of jobs to launch in parallel"""
 
 @dataclass
 class FeatureExtractionConfig:
-    data_dir: str
-    pretrained_model_path: str
-    model_name: str = "hiera_base_128x224"
-    output_dir: str = "features_outputs"
+    data_dir: str = MISSING
+    """Folder on disk containing manifests and videos"""
+
+    output_dir: str = MISSING
+    """Folder on disk where features will be saved"""
+
+    model_name: str = MISSING
+    """Hiera model identifier, e.g. hiera_base_128x224"""
+
+    pretrained_model_path: str = MISSING
+    """Path to trained feature extractor weights"""
+
     from_clip: bool = False
+    """Whether to initialize weights from a trained CLIP model"""
+
     split: str = "test"
+    """Dataset split, i.e. train, val or test"""
+
     fp16: bool = False
+    """Whether to extract features in fp16 rather than fp32"""
+
     do_aug: bool = False
+    """Whether to perform data augmentation"""
+
     epochs: int = 1
+    """Number of epochs to extract (useful when performing augmentation)"""
+
     num_frames: int = 128
+    """Number of video frames per clip"""
+
     sampling_rate: int = 2
+    """Sample every n-th frame"""
+
     target_fps: int = 25
+    """Target framerate in frames per second"""
+
     num_items_per_shard: int = 50
+    """Number of videos that will be processed sequentially in each shard"""
+
     max_batch_size: int = 2
+    """Maximum batch size (in 128-frame clips). Videos longer than this will be processed in shards."""
+
     video_backend: str = "pyav"
+    """Video backend. Should match the video backend used during pretraining."""
+
     launcher: LauncherConfig = LauncherConfig()
+    """Stopes launcher configuration"""
 
 def shard_generator(data: Any, shard_size: int) -> Generator[Any, None, None]:
     for i in range(0, len(data), shard_size):
@@ -70,6 +107,7 @@ class FeatureExtractionModule(StopesModule):
         Path(self.config.output_dir).mkdir(exist_ok=True, parents=True)
 
     def requirements(self) -> Requirements:
+        """Requirements for each submitted job"""
         return Requirements(
             nodes=1,
             mem_gb=6,
@@ -80,16 +118,24 @@ class FeatureExtractionModule(StopesModule):
         )
 
     def name(self) -> str:
-        return (
+        name = (
             f"feature_extractor_{self.config.split}_{self.config.model_name}_"
             f"{Path(self.config.pretrained_model_path).stem}"
         )
+        return name
 
     @property
     def num_shards(self) -> int:
         return math.ceil(self.num_items / self.config.num_items_per_shard)
 
+    def array(self) -> List[Any]:
+        return [1] * self.num_shards
+
     def load_model(self, device: Union[torch.device, str] = "cuda") -> SignHiera:
+        """
+        Loads a pretrained SignHiera model for feature extraction and moves it to specified device
+        """
+
         if self.config.from_clip:
             if "hiera" in self.config.model_name:
                 model = SignHiera.from_clip_model(
@@ -104,6 +150,7 @@ class FeatureExtractionModule(StopesModule):
             misc.load_model(model, self.config.pretrained_model_path)
 
         model.head = nn.Identity()
+
         print(f"Number of parameters: {sum([p.numel() for p in model.parameters()])}")
 
         model.eval()
@@ -112,6 +159,14 @@ class FeatureExtractionModule(StopesModule):
         return model
 
     def get_dataloader(self, indices: Tuple[int, int]) -> DataLoader:
+        """
+        Creates a dataloader for video feature extraction
+        Uses a fixed T//2 stride for the sliding window
+        """
+
+        if self.config.video_backend == "cuda":
+            torch.multiprocessing.set_start_method("spawn")
+
         dataset = VideoDataset(
             mode=self.config.split,
             video_backend=self.config.video_backend,
@@ -128,33 +183,40 @@ class FeatureExtractionModule(StopesModule):
             gpu=(torch.cuda.current_device() if self.config.video_backend == "cuda" else None),
         )
 
-        return DataLoader(          
-          dataset,
-          batch_size=1,
-          num_workers=1 if self.config.video_backend == "cuda" else 2,
-          persistent_workers=True,
-          pin_memory=not self.config.video_backend == "cuda",
+        dataloader = DataLoader(
+            dataset,
+            batch_size=1,
+            num_workers=1 if self.config.video_backend == "cuda" else 2,
+            persistent_workers=True,
+            pin_memory=not self.config.video_backend == "cuda",
         )
 
-    def run(self, iteration_value: Any, iteration_index: int):
+        return dataloader
+
+    def run(self, iteration_value: Any, iteration_index: int) -> None:
+        """
+        Extract features for a shard of the data
+        """
         output_dir = Path(self.config.output_dir)
+
         start_id, end_id = (
             self.config.num_items_per_shard * iteration_index,
             self.config.num_items_per_shard * (iteration_index + 1),
         )
-
         print(f"Indices: {start_id}-{end_id}")
+
         device = "cuda" if torch.cuda.is_available() else "cpu"
         print(f"Using device {device}")
 
         model = self.load_model(device=device)
         dataloader = self.get_dataloader(indices=(start_id, end_id))
 
+        # Split data into folders using prefix based on their video name
+        # Helps reduce load when file system is NFS
         fids = [
             Path(dataloader.dataset.path_to_videos[i]).stem for i in range(len(dataloader.dataset))
         ]
         prefixes = [fid[:5] for fid in fids]
-
         for epoch in tqdm(range(self.config.epochs), desc="Creating folder structure"):
             for prefix in prefixes:
                 prefix_path = output_dir / str(epoch) / prefix
@@ -179,12 +241,14 @@ class FeatureExtractionModule(StopesModule):
                 if padding.dim() == 2:
                     padding = rearrange(padding, "b r -> (b r)")
 
+                # Process in chunks
                 if len(frames) > self.config.max_batch_size:
                     shard_outputs = []
                     frames_shards = shard_generator(frames, self.config.max_batch_size)
                     padding_shards = shard_generator(padding, self.config.max_batch_size)
 
                     for frames_shard, padding_shard in zip(frames_shards, padding_shards):
+
                         with torch.inference_mode(), torch.cuda.amp.autocast(
                             enabled=self.config.fp16
                         ):
@@ -204,6 +268,7 @@ class FeatureExtractionModule(StopesModule):
                 fid = fids[idx]
                 prefix = prefixes[idx]
 
+                # Save to disk
                 output_file = output_dir / str(epoch) / prefix / f"{fid}.pt"
                 torch.save(outputs, output_file)
