+++ translation/utils/download_model.py
@@ -12,8 +12,10 @@ def download_model(url: str, save_dir: str = "."):
     model_path = Path(save_dir) / model_name
     if model_path.exists():
         print(f"Model already exists at: {model_path}")
+        return model_path
     else:
         print("Model not found, downloading...")
         filename = wget.download(url, str(model_path))
         print(f"Downloaded model to: {filename}")
+        return filename
